requirements.txt
opencv-python
numpy
tkinter
Pillow
imageio